+++ CHANGES.rst
@@ -4,8 +4,10 @@ Changelog
 5.0a1 (unreleased)
 ------------------
 
-- Don't install Archetypes-related packages.
-  Install plone.app.contenttypes.
+- Make PLONE_FIXTURE not install a content type system.
+  Packages that need content types to run their tests should
+  pick the appropriate fixture from plone.app.contenttypes
+  or Products.ATContentTypes.
   [davisagli]
 
 - Pin [robot] extra to ``robotsuite>=1.4.0``.